+++ src/module-loader.js
@@ -25,6 +25,7 @@ import flrtHelper from '@/flrt-helper';
 import hoverProfiles from '@/hover-profiles';
 import imageUpscaling from '@/image-upscaling';
 import inlineWiki from '@/inline-wiki';
+import inventoryLockAndHide from '@/inventory-lock-and-hide';
 import keyboardShortcuts from '@/keyboard-shortcuts';
 import lgsReminder from '@/lgs-reminder';
 import locationCatchStats from '@/location-catch-stats';
@@ -96,6 +97,7 @@ export default [
   hoverProfiles,
   imageUpscaling,
   inlineWiki,
+  inventoryLockAndHide,
   keyboardShortcuts,
   lgsReminder,
   locationCatchStats,